feat: add categories management with create, edit, and delete functionality
- Implemented CategoriesClient component for managing categories.
- Added CategoryModal for creating and editing categories.
- Integrated delete functionality with error handling.
- Created actions for user profile management including updating profile and changing password.
- Developed SettingsClient for user settings, including avatar selection and password management.
- Introduced TransactionsClient for managing transactions with filtering options.
- Added user avatar component with support for custom icons and images.
- Seeded default categories for new users.
- Created utility functions for avatar icons and default categories.
- Updated database schema to fix budget index issue.

+++ lib/validations.ts
@@ -3,7 +3,7 @@ import { ICON_OPTIONS, COLOR_OPTIONS } from "./icon-map";
 
 export const transactionSchema = z.object({
   type: z.enum(["income", "expense"]),
-  amount: z.coerce.number().positive().max(999_999_999),
+  amount: z.coerce.number().int().positive().max(999_999_999),
   categoryId: z.string().min(1, "Selecciona una categoría"),
   date: z.coerce.date(),
   description: z.string().trim().max(200).optional(),
